feat: add listCommentsOfCurrentUser in UserController and CommentService

+++ src/main/java/com/wesleybertipaglia/blog/service/CommentService.java
@@ -51,9 +51,12 @@ public class CommentService {
     }
 
     @Transactional(readOnly = true)
-    public Page<CommentResponseDTO> listCommentsByUser(int page, int size, UUID userId) {
+    public Page<CommentResponseDTO> listCommentsByUser(int page, int size, String tokenSubject) {
         Pageable pageable = PageRequest.of(page, size);
-        return commentRepository.findByUserId(userId, pageable).map(CommentMapper::convertToDTO);
+        User user = userRepository.findById(UUID.fromString(tokenSubject))
+                .orElseThrow(() -> new EntityNotFoundException("User not found"));
+
+        return commentRepository.findByUserId(user.getId(), pageable).map(CommentMapper::convertToDTO);
     }
 
     @Transactional(readOnly = true)
@@ -65,6 +68,12 @@ public class CommentService {
     public Optional<CommentResponseDTO> updateComment(UUID id, CommentCreateDTO commentCreateDTO, String tokenSubject) {
         Comment comment = commentRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Comment not found"));
+        User user = userRepository.findById(UUID.fromString(tokenSubject))
+                .orElseThrow(() -> new EntityNotFoundException("User not found"));
+
+        if (!comment.getUser().getId().equals(user.getId())) {
+            throw new IllegalArgumentException("User is not the owner of the comment");
+        }
 
         comment.setContent(commentCreateDTO.content());
         return Optional.of(CommentMapper.convertToDTO(commentRepository.save(comment)));
@@ -74,6 +83,13 @@ public class CommentService {
     public void deleteComment(UUID id, String tokenSubject) {
         Comment comment = commentRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Comment not found"));
+
+        User user = userRepository.findById(UUID.fromString(tokenSubject))
+                .orElseThrow(() -> new EntityNotFoundException("User not found"));
+
+        if (!comment.getUser().getId().equals(user.getId())) {
+            throw new IllegalArgumentException("User is not the owner of the comment");
+        }
 
         commentRepository.delete(comment);
     }
